+++ src/main/java/com/rolas/studies/service/topic/TopicServiceImpl.java
@@ -4,36 +4,38 @@ import java.util.Date;
 import java.util.List;
 
 import javax.inject.Inject;
-import javax.ws.rs.core.Response;
-import javax.ws.rs.core.UriBuilder;
-import javax.ws.rs.core.Response.Status;
-
 import com.rolas.studies.dao.topic.TopicDao;
 import com.rolas.studies.entities.Topic;
+import com.rolas.studies.entities.User;
+import com.rolas.studies.views.TopicVW;
 
 public class TopicServiceImpl implements TopicService {
 
 	@Inject
 	TopicDao topicDao;
 
-	
 	@Override
 	public Topic getTopic(Integer id) {
 		return topicDao.get(id);
 	}
-	
+
 	@Override
-	public List<Topic> getByCategory(Integer categoryId) {
-		return topicDao.getByCategory(categoryId);
+	public List<TopicVW> getByCategory(Integer categoryId) {
+		return categoryId != null ?topicDao.getByCategory(categoryId) : null;
 	}
 
 	@Override
-	public boolean delete(Integer id) {
-		return topicDao.delete(id);
+	public boolean delete(Integer id, User u) {
+		if (u.getUserName().equals(topicDao.get(id).getUser().getUserName()) || u.isAdmin()) {
+			return topicDao.delete(id);
+		} else {
+			return false;
+		}
 	}
 
 	@Override
 	public Topic insert(Topic topic) {
+		if(topic == null) return null;
 		topic.setInsertDate(new Date());
 		topic.setPostCount(0);
 		return topicDao.persist(topic);
@@ -42,11 +44,19 @@ public class TopicServiceImpl implements TopicService {
 	@Override
 	public boolean update(Topic topic) {
 		Topic fromDatabase = (Topic) topicDao.get(topic.getId());
-		fromDatabase.setDescription(topic.getDescription());
-		fromDatabase.setName(topic.getName());
-		fromDatabase.setAnswered(topic.isAnswered());
-		return topicDao.update(fromDatabase);
+		if (fromDatabase != null) {
+			fromDatabase.setDescription(topic.getDescription());
+			fromDatabase.setName(topic.getName());
+			fromDatabase.setAnswered(topic.isAnswered());
+			fromDatabase.setUpdateDate(new Date());
+			return topicDao.update(fromDatabase);
+		}
+		return false;
 	}
-	
-	
+
+	@Override
+	public List<TopicVW> getAll() {
+		return topicDao.getAll();
+	}
+
 }
